requirements.txt
numpy==1.24.3
Pillow==10.0.0
tkinter
matplotlib==3.7.2
cryptography==41.0.3
scipy==1.11.1
